clean: Move existing rpcv8 adapters to dedicated file & return by value instead of pointers when possible
