+++ rpc/v8/trace.go
@@ -65,6 +65,13 @@ type FunctionInvocation struct {
 	IsReverted         bool                         `json:"is_reverted"`
 }
 
+// https://github.com/starkware-libs/starknet-specs/blob/v0.3.0/api/starknet_api_openrpc.json#L2344
+type FunctionCall struct {
+	ContractAddress    felt.Felt   `json:"contract_address"`
+	EntryPointSelector felt.Felt   `json:"entry_point_selector"`
+	Calldata           []felt.Felt `json:"calldata"`
+}
+
 /****************************************************
 		Tracing Handlers
 *****************************************************/
